Crate twm_core

Re-exports

pub use config::toml;

Modules

bbox
Bounding boxes are used to hold the position and size of windows inside the tiling window manager.

config

display

layout
A layout is responsible for laying out tiles inside a workspace.

manager
The Manager is responsible for managing all workspaces and keeping track of the active.

tile
Tiles are used to move around and resize windows inside a workspace.

util
Contains model-independent enums, structs and traits.

window
Windows are used to represent running programs of a user.

workspace

Traits

DynClone
This trait is implemented by any type that implements std::clone::Clone.