Module twm_core::bbox

Bounding boxes are used to hold the position and size of windows inside the tiling window manager.

Structs

BBox
A bounding box.

BBoxBuilder
A builder for creating bounding boxes.

HorizontalSplit
The result of an horizontal split.

VerticalSplit
The result of an vertical split.

Enums

SplitDirection
The direction a split can occur.