Module twm_core::layout

A layout is responsible for laying out tiles inside a workspace.

These is done by changing the bounding box of each tile according to the layout's implementation details.

Modules

middle_layout

sided_layout
A sided layout uses have of the workspace's bounding box for a focused tile. The rest of the space is evenly shared between the rest of the tiles.

Structs

LayoutMeta
Metadata associated with a layout.

LayoutUpdateInfo
The information needed to update tiles inside a workspace.

Traits

DynClone
This trait is implemented by any type that implements std::clone::Clone.

Layout
A layout.